Adding new action into CMake.
