+++ src/Application/LayerIO/Actions/ActionExportPoints.cc
@@ -34,7 +34,7 @@
 #include <fstream>
 
 // Application includes
-#include <Application/LayerIO/Actions/ActionExportVector.h>
+#include <Application/LayerIO/Actions/ActionExportPoints.h>
 #include <Application/LayerIO/LayerIO.h>
 #include <Application/Layer/LayerManager.h>
 #include <Application/ProjectManager/ProjectManager.h>
@@ -43,26 +43,11 @@
 // Define a function that registers the action. The action also needs to be
 // registered in the CMake file.
 CORE_REGISTER_ACTION( Seg3D, ExportPoints )
-CORE_REGISTER_ACTION( Seg3D, ExportMatrix )
 
 namespace Seg3D
 {
-
   void PointWriter::operator()(std::ostream& o, const Core::Point& p) const
   {
     o << p.x() << " " << p.y() << " " << p.z() << std::endl;
   }
-
-  void MatrixDoubleWriter::operator()(std::ostream& o, const double& x) 
-  {
-    index_++;
-    o << x << "\t";
-    if (index_ % dim2_ == 0)
-    {
-      o << std::endl;
-    }
-  }
-
-  //template class ActionExportVector<Core::Point, PointWriter>;
-
 } // end namespace Seg3D
