+++ src/Application/LayerIO/Actions/ActionExportVector.h
@@ -141,87 +141,6 @@ private:
   WriterT writer_;
 };
 
-struct PointWriter
-{
-  void operator()(std::ostream& o, const Core::Point& p) const;
-};
-
-
-class ActionExportPoints : public ActionExportVector<Core::Point, PointWriter>
-{
-  CORE_ACTION(
-    CORE_ACTION_TYPE("ExportPoints", "This action exports a list of points (x y z) to file.")
-    CORE_ACTION_ARGUMENT("file_path", "Path to the file that the points will be written to.")
-    CORE_ACTION_ARGUMENT("points", "3D points.")
-    CORE_ACTION_CHANGES_PROJECT_DATA()
-    )
-
-    
-public:
-  using Base = ActionExportVector<Core::Point, PointWriter>;
-  ActionExportPoints(const std::string& file_path, const std::vector<Core::Point>& points) :
-    Base(file_path, points, PointWriter())
-  {
-    init_parameters();
-  }
-  ActionExportPoints()
-  {
-    init_parameters();
-  }
-
-  // DISPATCH:
-  static void Dispatch(Core::ActionContextHandle context,
-    const std::string& file_path,
-    const std::vector<Core::Point>& points)
-  {
-    auto action = new ActionExportPoints(file_path, points);
-    Core::ActionDispatcher::PostAction(Core::ActionHandle(action), context);
-  }
-};
-
-class MatrixDoubleWriter
-{
-public:
-  MatrixDoubleWriter() {}
-  MatrixDoubleWriter(int dim1, int dim2) : dim1_(dim1), dim2_(dim2) {}
-  void operator()(std::ostream& o, const double& x);
-private:
-  int dim1_{ 0 }, dim2_{ 0 };
-  int index_{ 0 };
-};
-
-class ActionExportMatrix : public ActionExportVector<double, MatrixDoubleWriter>
-{
-  CORE_ACTION(
-    CORE_ACTION_TYPE("ExportMatrix", "This action exports a maxtrix of a given size.")
-    CORE_ACTION_ARGUMENT("file_path", "Path to the file that the matrix will be written to.")
-    CORE_ACTION_ARGUMENT("matrix", "matrix")
-    CORE_ACTION_CHANGES_PROJECT_DATA()
-    )
-
-
-public:
-  using Base = ActionExportVector<double, MatrixDoubleWriter>;
-  ActionExportMatrix(const std::string& file_path, const std::vector<double>& matrix, int dim1, int dim2) :
-    Base(file_path, matrix, MatrixDoubleWriter(dim1, dim2))
-  {
-    init_parameters();
-  }
-  ActionExportMatrix()
-  {
-    init_parameters();
-  }
-
-  // DISPATCH:
-  static void Dispatch(Core::ActionContextHandle context,
-    const std::string& file_path,
-    const std::vector<double>& matrix, int dim1, int dim2)
-  {
-    auto action = new ActionExportMatrix(file_path, matrix, dim1, dim2);
-    Core::ActionDispatcher::PostAction(Core::ActionHandle(action), context);
-  }
-};
-
 } // end namespace Seg3D
 
 #endif
